move loopback view transform to own file
