+++ src/v3/src/transformer/terminal/odaEnrollment/index.ts
@@ -11,3 +11,5 @@
  */
 
 export * from './transformOdaEnrollment';
+export * from './transformOdaEnrollmentAndroidAppLink';
+export * from './transformOdaEnrollmentLoopback';
